requirements.txt
# CDE API
fastapi>=0.115
uvicorn[standard]>=0.32
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
psycopg2-binary>=2.9
pydantic-settings>=2.0
python-multipart>=0.0.9

# Agent Pipeline
langgraph>=0.4
langchain>=0.3
langchain-openai>=0.3
langchain-community>=0.3
langchain-qdrant>=0.2
qdrant-client>=1.12
fastembed>=0.4
httpx>=0.27
python-dotenv>=1.0
chainlit>=1.1

# IFC Processing
ifcopenshell>=0.8

# Data Processing
pandas>=2.0
openpyxl>=3.1

# Chunking
docling>=2.0
